requirements.txt
absl-py==2.1.0
ansicon==1.89.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
blessed==1.20.0
blinker==1.8.1
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cors==1.0.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
editor==1.6.6
executing==2.0.1
Faker==25.1.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
gevent==24.2.1
glfw==2.6.5
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h2o-wave==1.2.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
inquirer==3.2.4
intel-tensorflow==0.0.1
ipykernel==6.29.2
ipython==8.22.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jinxed==1.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.10.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pygame==2.5.2
Pygments==2.17.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
readchar==4.0.6
requests==2.31.0
requests-file==2.1.0
rich==13.7.1
runs==1.2.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tldextract==5.1.2
torch==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
xmod==1.8.1
zope.event==5.0
zope.interface==7.0.1
